4.21 delete with token

+++ bloglist/controllers/blogs.js
@@ -19,7 +19,7 @@ blogsRouter.post('/', async (req, res, next) => {
     try{
         const decodedToken = jwt.verify(req.token, process.env.SECRET);
     
-        if (!token || !decodedToken.id){
+        if (!decodedToken.id){
             return res.status(401).json({error:'token missing or invalid'});
         }
 
@@ -41,7 +41,18 @@ blogsRouter.post('/', async (req, res, next) => {
 
 blogsRouter.delete('/:id', async (req, res, next) => {
     try{
-        const r = await Blog.deleteOne({_id:req.params.id})
+        const decodedToken = jwt.verify(req.token, process.env.SECRET);
+    
+        if (!decodedToken.id){
+            return res.status(401).json({error:'token missing or invalid'});
+        }
+
+        const blog = await Blog.findOne({_id:req.params.id})
+        if (blog.user.toString() != decodedToken.id){
+            return res.status(401).json({error:'token invalid to delete'});
+        }
+
+        await Blog.deleteOne({_id:req.params.id})
         res.status(204).send();
     }
     catch(e){
